move GlobalRSI next to RSI
