+++ cpp/src/main/models/GlobalRankingScoreIterator.cpp
